начал разрабатывать ТЗ-10. Где-то 2/3 - многослойность сделал, сервисный слой, экзепшн-хендлер

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -15,5 +15,5 @@ public class Film {
     private String description;
     private LocalDate releaseDate;
     private Duration duration;
-    private Set<Long> likes;
+    private Set<Integer> likes;
 }
